pipeline1/requirements.txt
openai>=1.0.0
pydantic>=2.0.0
instructor>=0.4.0
python-dotenv>=1.0.0
jsonlines>=4.0.0
tqdm>=4.66.0
google-genai>=0.7.0
httpx>=0.24.0
aiohttp>=3.8.0 